pyproject.toml
[tool.poetry]
name = "finlearn-api"
version = "0.1.0"
description = ""
authors = ["Leonardo Magalhães <leonardo@biomainvestimentos.com>"]
readme = "README.md"
packages = [{include = "finlearn_api"}]

[tool.poetry.dependencies]
python = "^3.11.2"
alembic = "^1.10.2"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
click = "^8.1.3"
sqlalchemy = "^2.0.7"
pydantic = "^1.10.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
